backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
stripe>=7.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
